Art and resources
Moved a large number of sprites, both item and skill, across to a new home. As such, all new sprites have the working files inside "Art" and their icon arts are added to "Resources". Edits made to the alchemy icon. New changes added to databases as well. This is the art/data pass for today's code.