Code update & version upgrade
Version upgrade to .26 of Unity 2019.4, code has been tweaked to allow for icon artwork to be handled by resources. Looking at the possibility for putting the databases inside of Resources later. New tag designed for handling items when they collide with a phantom circle at a tbd point in time.