Fixed movement and changes stuff up
The player movement script has been updated and is no longer acting on velocity, instead it's doing some good ol' rigidBody.MovePosition() function calls. Splitting up movement and input to not cause a complete input collapse, added a function that handles Time.DeltaTime and a function to handle the conversion to Input (might change this later).